+++ client/src/config.rs
@@ -66,6 +66,8 @@ pub struct RuntimeConfig {
 	pub tracking_service_ping_interval: u64,
 	/// Don't update light client if update is available (default: false)
 	pub no_update: bool,
+	/// Perform random maintenance restart
+	pub maintenance_restart: bool,
 }
 
 impl From<&RuntimeConfig> for SyncClientConfig {
@@ -139,6 +141,7 @@ impl Default for RuntimeConfig {
 			tracking_service_address: "http://127.0.0.1:8989".to_string(),
 			tracking_service_ping_interval: 10,
 			no_update: false,
+			maintenance_restart: false,
 		}
 	}
 }